Fixes #67 Shows infoDialog when podman preset is used
Currently we do not support the podman preset, 'focus' on OpenShift
and MicroShift, as decided at the F2F. This means we need to tell
the user we would have to change the preset to work correctly.
The current text is temporary until we have a better solution, like
markdown formatted or decided to implement support.