chore: run crc installer with full user experience
previously the installer was executed as a hidden process and did not show any option to reboot the computer even when it was required. Then the flow keep processing and the setup ended up on an error due to not being able to stop it or to reboot the computer.

Also this change makes the openshift local installer matches how other installers are executed showing full experience to the user (i.e. podman installer)

Signed-off-by: Adrian Riobo Lorenzo <ariobolo@redhat.com>